added vm support

+++ wrapper_out.py
@@ -2,16 +2,7 @@ import subprocess
 import multiprocessing
 import os
 import time
-
-# List of URLs to be crawled
-urls = open('websites1.txt', 'r').read().splitlines()
-
-# Docker container names
-containers = ["accads_control", "accads_adblock"]
-# containers = ["accads_control"]
-
-# Build Docker images (assuming Dockerfiles are in the current directory)
-subprocess.run(["docker", "build", "-t", "accads", "-f", "Dockerfile", "."])
+import argparse
 
 # Function to create a directory with 777 permission
 def create_directory(dir_name):
@@ -21,6 +12,8 @@ def create_directory(dir_name):
         print(f"Created {dir_name} with 777 permissions.")
     else:
         print(f"{dir_name} already exists.")
+        print('CLEANING THE DIRECTORY OF ANY FILES')
+        os.system(f'rm -rf ./{dir_name}/*')
 
 def create_data_directories():
     # Directories to be created
@@ -49,31 +42,61 @@ def handle_container(container_name, image_name, url, extn):
     feed_url_to_container(container_name, url, extn)
     time.sleep(1)  # Add delay if necessary
 
+# List of URLs to be crawled
+urls = open('websites1.txt', 'r').read().splitlines()
+docker = 0
+vm = 0
+parser = argparse.ArgumentParser(description='Specify Extension for wrapper_out.py')
+parser.add_argument('--extn', type=str)
+args = parser.parse_args()
+
 create_data_directories()
-for url in urls:
-    # Create multiprocessing processes
-    processes = []
-    image_name = "accads"
-    p1 = multiprocessing.Process(target=handle_container, args=(containers[0], image_name, url, 'control'))
-    p2 = multiprocessing.Process(target=handle_container, args=(containers[1], image_name, url, 'adblock'))
-    
-    p1.start()
-    p2.start()
 
-    TIMEOUT = 70
-    start = time.time()
-    print("joining jobs")
-    # Wait for all processes to finish
-    p1.join(timeout = 60)
-    p2.join(timeout = 60)
+if docker:
+    # Docker container names
+    containers = ["accads_control", "accads_adblock"]
+    # containers = ["accads_control"]
 
-    while time.time() - start <= TIMEOUT:
-        if p1.is_alive():
-            p1.terminate()
-        if p2.is_alive():
-            p2.terminate()
+    # Build Docker images (assuming Dockerfiles are in the current directory)
+    subprocess.run(["docker", "build", "-t", "accads", "-f", "Dockerfile", "."])
+    for url in urls:
+        # Create multiprocessing processes
+        processes = []
+        image_name = "accads"
+        p1 = multiprocessing.Process(target=handle_container, args=(containers[0], image_name, url, 'control'))
+        p2 = multiprocessing.Process(target=handle_container, args=(containers[1], image_name, url, 'adblock'))
+        
+        p1.start()
+        p2.start()
 
-    time.sleep(2)
+        TIMEOUT = 70
+        start = time.time()
+        print("joining jobs")
+        # Wait for all processes to finish
+        p1.join(timeout = 60)
+        p2.join(timeout = 60)
+
+        while time.time() - start <= TIMEOUT:
+            if p1.is_alive():
+                p1.terminate()
+            if p2.is_alive():
+                p2.terminate()
+
+        time.sleep(2)
+
+elif vm:
+    for url in urls:
+        try:
+            # Execute a command with a timeout of 5 seconds
+            result = subprocess.run(['python3', 'wrapper_in.py', '--url', url, '--extn', args.extn], timeout=220)
+            print("Command completed:", result)
+        except subprocess.TimeoutExpired:
+            print("Command timed out and was terminated.")
+
+        os.system(f'python3 wrapper_in.py --url {url} --extn {args.extn}')
+else:
+    print('PLEASE SPECIFY EITHER OF DOCKER OR VM')
+
 
 # Optionally, stop and remove containers after use
 # for container in containers:
